feat: remove AWS components, move interface of IdGenerator and add its web version
