+++ packages/opentelemetry-api/src/index.ts
@@ -51,7 +51,6 @@ export * from './trace/trace_flags';
 export * from './trace/trace_state';
 export * from './trace/tracer_provider';
 export * from './trace/tracer';
-export * from './platform/node/IdGenerator';
 
 export { Context } from '@opentelemetry/context-base';
