+++ packages/opentelemetry-api/src/platform/node/index.ts
@@ -15,4 +15,3 @@
  */
 
 export * from './globalThis';
-export * from './IdGenerator';
