+++ packages/opentelemetry-core/src/platform/browser/index.ts
@@ -17,7 +17,7 @@
 export * from './BasePlugin';
 export * from './environment';
 export * from './hex-to-base64';
-export * from './id';
+export * from './RandomIdGenerator';
 export * from './performance';
 export * from './sdk-info';
 export * from './timer-util';
