+++ packages/opentelemetry-core/src/platform/node/index.ts
@@ -21,4 +21,3 @@ export * from './RandomIdGenerator';
 export * from './performance';
 export * from './sdk-info';
 export * from './timer-util';
-export * from './AWSXrayIdGenerator';
